+++ src/ros/cameras/camera_msgs/msg/Camera.msg
@@ -1,3 +1,4 @@
 string serial
 string node
 string original_serial
+string path
