+++ src/ros/cameras/cameras3/include/cameras/globals.hpp
