+++ src/ros/cameras/cameras3/include/properties/common.hpp
@@ -2,7 +2,9 @@
 #define COMMON_PROPERTY_HEADER
 
 #include <string>
+#include <stdlib.h>
 #include <gst/gst.h>
+#include "pipelines/properties.hpp"
 
 bool link_elements(rclcpp::Node* streamer_node, GstElement* first_element, GstElement* second_element, const std::string serial);
 
@@ -13,5 +15,7 @@ int set_property(rclcpp::Node* streamer_node, const std::string serial, const st
 bool set_property(rclcpp::Node* streamer_node, const std::string serial, const std::string profile, const std::string original_serial, const std::string element, const bool default_value);
 
 bool verify_resolution(const std::string device_name, std::string* mime, int* width, int* height, int* framerate, int* framerate_denominator);
+
+void match_lost_devname(rclcpp::Node* streamer_node, std::unordered_map<std::string, Pipeline*>* pipelines);
 
 #endif
