--- src/ros/cameras/cameras3/params/.gitignore
@@ -1,1 +1,0 @@
-/dev
