+++ src/ros/cameras/cameras3/params/dev/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
