--- src/ros/cameras/cameras3/src/camera_pipelines.cpp
@@ -1,888 +1,0 @@
-#include <string>
-
-#include <gst/gst.h>
-#include "rclcpp/rclcpp.hpp"
-#include <camera_msgs/msg/camera.hpp>
-
-#include "cameras/pipeline.hpp"
-
-#include "properties/common.hpp"
-#include "properties/sources.hpp"
-#include "properties/sinks.hpp"
-
-#include "properties/capsfilters.hpp"
-#include "properties/cpufilters.hpp"
-
-#include "properties/h264.hpp"
-#include "properties/vpX.hpp"
-#include "properties/av1.hpp"
-
-/*
- * V4l camera to webrtc pipeline
- * converts any v4l source to raw video and then encodes a format for webrtc
- * gst-launch-1.0 v4l2src device={props->node} ! {props->mime},width={props->width},height={props->height},framerate={props->framerate}/1 ! decodebin ! videoconvert ! webrtcsink meta='meta, serial=(string){props->serial}' video-caps={props->video_caps}
- */
-
-
-GstElement* v4l2webrtc_pipeline(rclcpp::Node* streamer_node, v4l2webrtcPipelineProperties* props)
-{
-  /* 
-     This creates a v4l2src to webrtc pipeline with the following structure:
-     v4l2src ! capsfilter ! decodebin3 ! videoconvertscale ! scalefilter ! (clockoverlay) ! webrtcsink
-     @param streamer_node pointer to the ros2 streamer node
-     @param props pointer to the pipeline properties
-     @return GstElement* pointer to the created GStreamer pipeline
-  */
-
-  // 0. Initialize constants
-
-  // Verify resolution
-  const std::string pipeline_type = "v4l2webrtc";
-  if (verify_resolution(props->device, &props->mime, &props->width, &props->height, &props->framerate, &props->framerate_denominator)) {
-      RCLCPP_INFO(streamer_node->get_logger(), "Starting pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  } else {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Wrong resolution! Fallback %s pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  };
-
-  // Disable crop43 if it is already 4:3
-  const int crop_width = crop43(props->width, props->height);
-  if (crop_width == 0) {
-      props->crop43 = false;
-  }
-
-  // 1. Create the elements
-  GstElement* gst_pipeline = gst_pipeline_new(props->serial.c_str());
-  GstElement* source = gst_element_factory_make("v4l2src", "video-source");
-  GstElement* srcfilter = gst_element_factory_make("capsfilter", "srcfilter");
-  GstElement* rate = (props->downrate > 1) ? gst_element_factory_make("videorate", "rater") : nullptr;
-  GstElement* decode = gst_element_factory_make("decodebin3", "decoder");
-  GstElement* convert = gst_element_factory_make("videoconvertscale", "converter");
-  GstElement* scalefilter = gst_element_factory_make("capsfilter", "scalefilter");
-  GstElement* clock = props->show_clock ? gst_element_factory_make("clockoverlay", "clock") : nullptr;
-  GstElement* cropper = props->crop43 ? gst_element_factory_make("videocrop", "video-cropper") : nullptr;
-  GstElement* webrtc = gst_element_factory_make("webrtcsink", "webrtc");
-
-
-  if (!gst_pipeline || !source || (props->downrate > 1 && !rate) || !srcfilter || !decode || !convert || !scalefilter || (props->show_clock && !clock) || (props->crop43 && !cropper) || !webrtc 
-      ) {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Could not create pipeline for %s", props->serial.c_str());
-      return nullptr;
-  }
-
-  // 2. Set element properties
-  set_v4lsource(source, props->device, props->io_mode);
-  set_srcfilter(srcfilter, props->mime, props->width, props->height, props->framerate, props->framerate_denominator, props->downrate, props->brightness, props->contrast);
-  set_convertscale(convert, props->chroma_resampler, props->dither, props->method);
-  set_scalefilter(scalefilter, props->format, props->width, props->height, props->framerate, props->framerate_denominator, props->downscale, props->downrate, props->brightness, props->contrast);
-  set_crop43(cropper, props->crop43, crop_width, props->downscale);
-  set_webrtcsink(webrtc, props->serial, props->video_caps, props->do_fec, props->do_retransmission, props->congestion_control, props->bitrate);
-
-  // 3. Add elements to pipeline
-  gst_bin_add_many(GST_BIN(gst_pipeline), source, srcfilter, decode, convert, scalefilter, webrtc, NULL);
-  if (props->downrate > 1) gst_bin_add(GST_BIN(gst_pipeline), rate);
-  if (props->crop43) gst_bin_add(GST_BIN(gst_pipeline), cropper);
-  if (props->show_clock) gst_bin_add(GST_BIN(gst_pipeline), clock);
-
-  // 4. Link elements
-  
-  // Change fps
-  if (props->downrate > 1) {
-    if (!link_elements(streamer_node, source, rate, props->serial)) return nullptr;
-    if (!link_elements(streamer_node, rate, srcfilter, props->serial)) return nullptr;
-  } else {
-    if (!link_elements(streamer_node, source, srcfilter, props->serial)) return nullptr;
-  }
-
-  // Convert to raw
-  if (!link_elements(streamer_node, srcfilter, decode, props->serial)) return nullptr;
-
-  g_signal_connect(decode, "pad-added", G_CALLBACK(+[](GstElement* , GstPad* new_pad, gpointer user_data) {
-      GstElement* convert = static_cast<GstElement*>(user_data);
-      GstPad* sink_pad = gst_element_get_static_pad(convert, "sink");
-      if (sink_pad && !gst_pad_is_linked(sink_pad)) {
-          gst_pad_link(new_pad, sink_pad);
-      }
-      if (sink_pad) gst_object_unref(sink_pad);
-  }), convert);
-
-  if (!link_elements(streamer_node, convert, scalefilter, props->serial)) return nullptr;
-
-  // Enable crop and/or clock
-  if (props->crop43 && props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, webrtc, props->serial)) return nullptr;
-  } else
-  if (props->crop43) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, webrtc, props->serial)) return nullptr;
-  } else
-  if (props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, webrtc, props->serial)) return nullptr;  } 
-  else {
-      if (!link_elements(streamer_node, scalefilter, webrtc, props->serial)) return nullptr;
-  }
-
-  return gst_pipeline;
-}
-
-
-/*
- * Retrieve ros2 parameters for v4l2webrtc pipeline or sets defaults
-*/
-
-v4l2webrtcPipelineProperties* get_v4l2webrtc_pipeline_properties(rclcpp::Node* streamer_node, camera_msgs::msg::Camera* camera)
-{
-  /*
-    Pulls ros2 parameters for a given camera and returns a properties struct for the v4l2webrtc pipeline creation function.
-    @param streamer_node pointer to the ros2 streamer node
-    @param camera pointer to the camera message containing at least the serial and node for the camera
-    @return pointer to a v4l2webrtcPipelineProperties struct containing the properties for the pipeline
-  */
-
-  // 0. Initialize constants
-  v4l2webrtcPipelineProperties* props = new v4l2webrtcPipelineProperties;
-  RCLCPP_DEBUG(streamer_node->get_logger(), "Getting props for %s", camera->serial.c_str());
-  props->serial = camera->serial;
-  props->node = camera->node;
-  props->original_serial = camera->original_serial;
-
-  // Get profile
-  std::string profile = "NULL";
-  streamer_node->get_parameter_or<std::string>((std::string(PIPELINE_PREFIX) + "." + camera->serial + ".profile").c_str(), profile, profile);
-  streamer_node->get_parameter_or<std::string>((std::string(DEFAULT_PREFIX) + "." + camera->original_serial + ".profile").c_str(), profile, profile);
-
-  // 1. Define default properties
-  std::string default_string;
-
-  // source
-  props->device = set_property(streamer_node, camera->serial, profile, camera->original_serial, "device", props->node);
-  default_string = "mmap";
-  props->io_mode = set_property(streamer_node, camera->serial, profile, camera->original_serial, "io_mode", default_string);
-
-  // filter
-  default_string = "I420";
-  props->format = set_property(streamer_node, camera->serial, profile, camera->original_serial, "format", default_string);
-  default_string = "image/jpeg";
-  props->mime = set_property(streamer_node, camera->serial, profile, camera->original_serial, "mime", default_string);
-
-  props->brightness = set_property(streamer_node, camera->serial, profile, camera->original_serial, "brightness", 0);
-  props->contrast = set_property(streamer_node, camera->serial, profile, camera->original_serial, "contrast", 0);
-  props->framerate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate", 30);
-  props->framerate_denominator = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate_denominator", 1);
-  props->height = set_property(streamer_node, camera->serial, profile, camera->original_serial, "height", 720);
-  props->width = set_property(streamer_node, camera->serial, profile, camera->original_serial, "width", 1280);
-
-  // convert
-  default_string = "nearest";
-  props->chroma_resampler = set_property(streamer_node, camera->serial, profile, camera->original_serial, "chroma_resampler", default_string);
-  default_string = "bayer";
-  props->dither = set_property(streamer_node, camera->serial, profile, camera->original_serial, "dither", default_string);
-  default_string = "nearest-neighbour";
-  props->method = set_property(streamer_node, camera->serial, profile, camera->original_serial, "method", default_string);
-
-  // scale
-  props->downscale = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downscale", 1);
-
-  // rate
-  props->downrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downrate", 1);
-
-  // cropper
-  props->crop43 = set_property(streamer_node, camera->serial, profile, camera->original_serial, "crop43", false);
-
-  // clock
-  props->show_clock = set_property(streamer_node, camera->serial, profile, camera->original_serial, "show_clock", false);
-
-  // webrtc
-  default_string = "gcc";
-  props->congestion_control = set_property(streamer_node, camera->serial, profile, camera->original_serial, "congestion_control", default_string);
-  default_string = "video/x-h264,profile=constrained-baseline"; 
-  props->video_caps = set_property(streamer_node, camera->serial, profile, camera->original_serial, "video_caps", default_string);
-
-  props->bitrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "bitrate", 4096);
-
-  props->do_fec = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_fec", false);
-  props->do_retransmission = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_retransmission", false);
-
-  return props;
-}
-
-/*
- * V4l camera (h264) to webrtc pipeline (direct)
- * Enforces alignment from h264 v4l camera and feeds directly to webrtc 
- * gst-launch-1.0 v4l2src device={props->node} ! {props->mime},width={props->width},height={props->height},framerate={props->framerate}/1,alignment={props->alignment},stream-format={props->stream_format},format={props->format}! webrtcsink meta='meta, serial=(string){props->serial}' video-caps=video/x-h264
- */
-GstElement* h264passthrough_pipeline(rclcpp::Node* streamer_node, h264passthroughPipelineProperties* props)
-{
-  // 0. Initialize constants
-
-  // Verify resolution
-  const std::string pipeline_type = "h264passthrough";
-  if (verify_resolution(props->device, &props->mime, &props->width, &props->height, &props->framerate, &props->framerate_denominator)) {
-      RCLCPP_INFO(streamer_node->get_logger(), "Starting pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  } else {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Wrong resolution! Fallback %s pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  };
-
-
-  // 1. Create the elements
-  GstElement* gst_pipeline = gst_pipeline_new(props->serial.c_str());
-  GstElement* source = gst_element_factory_make("v4l2src", "video-source");
-  GstElement* srcfilter = gst_element_factory_make("capsfilter", "srcfilter");
-  GstElement* parse = gst_element_factory_make("h264parse", "parser");
-  GstElement* webrtc = gst_element_factory_make("webrtcsink", "webrtc");
-  GstElement* payload = (props->payload_quirk) ? gst_element_factory_make("rtph264pay", "payloader") : nullptr;
-  GstElement* depayload = (props->payload_quirk) ? gst_element_factory_make("rtph264depay", "depayloader") : nullptr;
-
-  if (!gst_pipeline || !source || !srcfilter || !parse || !webrtc || (props->payload_quirk && !payload) || (props->payload_quirk && !depayload)) {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Could not create pipeline for %s", props->serial.c_str());
-      return nullptr;
-  }
-
-  // 2. Set element properties
-  set_v4lsource(source, props->device, props->io_mode);
-  set_srcfilter(srcfilter, props->mime, props->width, props->height, props->framerate, props->framerate_denominator, props->downrate, props->brightness, props->contrast);
-  set_h264payload(payload, props->payload_quirk);
-  set_h264parse(parse);
-  set_webrtcsink(webrtc, props->serial, props->video_caps, props->do_fec, props->do_retransmission, props->congestion_control, props->bitrate);
-
-  // 3. Add elements to pipeline
-  gst_bin_add_many(GST_BIN(gst_pipeline), source, srcfilter, parse, webrtc, NULL);
-  if (props->payload_quirk) gst_bin_add_many(GST_BIN(gst_pipeline), payload, depayload, NULL);
-
-  // 4. Link elements
-  if (!link_elements(streamer_node, source, srcfilter, props->serial)) return nullptr;
-
-  if (props->payload_quirk) {
-    if (!link_elements(streamer_node, srcfilter, payload, props->serial)) return nullptr;
-    if (!link_elements(streamer_node, payload, depayload, props->serial)) return nullptr;
-    if (!link_elements(streamer_node, depayload, parse, props->serial)) return nullptr;
-  } else {
-    if (!link_elements(streamer_node, srcfilter, parse, props->serial)) return nullptr;
-  }
-  if (!link_elements(streamer_node, parse, webrtc, props->serial)) return nullptr;
-
-  return gst_pipeline;
-}
-
-
-/*
- * Retrieve ros2 parameters for h264passthrough pipeline or sets defaults
-*/
-
-h264passthroughPipelineProperties* get_h264passthrough_pipeline_properties(rclcpp::Node* streamer_node, camera_msgs::msg::Camera* camera)
-{
-  // 0. Initialize constants
-  h264passthroughPipelineProperties* props = new h264passthroughPipelineProperties;
-  RCLCPP_DEBUG(streamer_node->get_logger(), "Getting props for %s", camera->serial.c_str());
-  props->serial = camera->serial;
-  props->node = camera->node;
-  props->original_serial = camera->original_serial;
-
-  // Get profile
-  std::string profile = "NULL";
-  streamer_node->get_parameter_or<std::string>((std::string(PIPELINE_PREFIX) + "." + camera->serial + ".profile").c_str(), profile, profile);
-  streamer_node->get_parameter_or<std::string>((std::string(DEFAULT_PREFIX) + "." + camera->original_serial + ".profile").c_str(), profile, profile);
-
-  // 1. Define default properties
-  std::string default_string;
-
-  // source
-  props->device = set_property(streamer_node, camera->serial, profile, camera->original_serial, "device", props->node);
-  props->io_mode = 4; // dmabuf
-
-  // rate
-  props->downrate = 1; // Do not change framerate
-
-  // filter
-  props->mime = "video/x-h264";
-
-  props->brightness = set_property(streamer_node, camera->serial, profile, camera->original_serial, "brightness", 0);
-  props->contrast = set_property(streamer_node, camera->serial, profile, camera->original_serial, "contrast", 0);
-  props->framerate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate", 30);
-  props->framerate_denominator = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate_denominator", 1);
-  props->height = set_property(streamer_node, camera->serial, profile, camera->original_serial, "height", 720);
-  props->width = set_property(streamer_node, camera->serial, profile, camera->original_serial, "width", 1280);
-
-  // payloader
-  props->payload_quirk = set_property(streamer_node, camera->serial, profile, camera->original_serial, "payload_quirk", false);
-
-  // webrtc
-  default_string = "gcc";
-  props->congestion_control = set_property(streamer_node, camera->serial, profile, camera->original_serial, "congestion_control", default_string);
-  props->video_caps = "video/x-h264";
-
-  props->bitrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "bitrate", 4096);
-
-  props->do_fec = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_fec", false);
-  props->do_retransmission = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_retransmission", false);
-
-  return props;
-}
-
-/*
- * V4l camera (any) decoded then encoded into x264enc
- * Enforces alignment from h264 v4l camera and feeds directly to webrtc 
- * gst-launch-1.0 v4l2src device={props->node} ! {props->mime},width={props->width},height={props->height},framerate={props->framerate}/1,alignment={props->alignment},stream-format={props->stream_format},format={props->format}! webrtcsink meta='meta, serial=(string){props->serial}' video-caps=video/x-h264
- */
-
-GstElement* h264software_pipeline(rclcpp::Node* streamer_node, h264softwarePipelineProperties* props)
-{
-  // 0. Initialize constants
-
-  // Verify resolution
-  const std::string pipeline_type = "h264software";
-  if (verify_resolution(props->device, &props->mime, &props->width, &props->height, &props->framerate, &props->framerate_denominator)) {
-      RCLCPP_INFO(streamer_node->get_logger(), "Starting pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  } else {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Wrong resolution! Fallback %s pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  };
-
-
-  // Disable crop43 if it is already 4:3 or jpeg
-  const int crop_width = crop43(props->width, props->height);
-  if (crop_width == 0) {
-      props->crop43 = false;
-  }
-
-  // 1. Create the elements
-  GstElement* gst_pipeline = gst_pipeline_new(props->serial.c_str());
-  GstElement* source = gst_element_factory_make("v4l2src", "video-source");
-  GstElement* rate = (props->downrate > 1) ? gst_element_factory_make("videorate", "rater") : nullptr;
-  GstElement* srcfilter = gst_element_factory_make("capsfilter", "srcfilter");
-  GstElement* decode = (props->mime == "image/jpeg") ? gst_element_factory_make(props->decoder.c_str(), "decoder") : nullptr;
-  GstElement* convert = gst_element_factory_make("videoconvertscale", "converter");
-  GstElement* scalefilter = gst_element_factory_make("capsfilter", "scalefilter");
-  GstElement* clock = (props->show_clock) ? gst_element_factory_make("clockoverlay", "clock") : nullptr;
-  GstElement* cropper = (props->crop43) ? gst_element_factory_make("videocrop", "video-cropper") : nullptr;
-  GstElement* encode = gst_element_factory_make("x264enc", "encoder");
-  GstElement* parse = gst_element_factory_make("h264parse", "parser");
-  GstElement* webrtc = gst_element_factory_make("webrtcsink", "webrtc");
-
-  if (!gst_pipeline || !source || (props->downrate > 1 && !rate) || !srcfilter || (props->mime == "image/jpeg" && !decode) || !convert || !scalefilter || (props->show_clock && !clock) || (props->crop43 && !cropper) || !encode || !parse || !webrtc) {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Could not create pipeline for %s", props->serial.c_str());
-      return nullptr;
-  }
-
-  // 2. Set element properties
-  set_v4lsource(source, props->device, props->io_mode);
-  set_srcfilter(srcfilter, props->mime, props->width, props->height, props->framerate, props->framerate_denominator, props->downrate, props->brightness, props->contrast);
-  set_convertscale(convert, props->chroma_resampler, props->dither, props->method);
-  set_scalefilter(scalefilter, props->format, props->width, props->height, props->framerate, props->framerate_denominator, props->downscale, props->downrate, props->brightness, props->contrast);
-  set_crop43(cropper, props->crop43, crop_width, props->downscale);
-  set_x264enc(encode, props->tune, props->speed_preset, props->threads, props->bitrate, props->noise_reduction, props->gop, props->framerate, props->framerate_denominator, props->downrate);
-  set_h264parse(parse, props->gop);
-  set_webrtcsink(webrtc, props->serial, props->video_caps, props->do_fec, props->do_retransmission, props->congestion_control, props->bitrate);
-
-  // 3. Add elements to pipeline
-  gst_bin_add_many(GST_BIN(gst_pipeline), source, srcfilter, convert, scalefilter, encode, parse, webrtc, NULL);
-  if (props->downrate > 1) gst_bin_add(GST_BIN(gst_pipeline), rate);
-  if (props->crop43) gst_bin_add(GST_BIN(gst_pipeline), cropper);
-  if (props->show_clock) gst_bin_add(GST_BIN(gst_pipeline), clock);
-  if (props->mime == "image/jpeg") gst_bin_add(GST_BIN(gst_pipeline), decode);
-
-  // 4. Link elements
-
-  // Change fps
-  if (props->downrate > 1) {
-    if (!link_elements(streamer_node, source, rate, props->serial)) return nullptr;
-    if (!link_elements(streamer_node, rate, srcfilter, props->serial)) return nullptr;
-  } else {
-    if (!link_elements(streamer_node, source, srcfilter, props->serial)) return nullptr;
-  }
-
-  // Convert to raw
-  if (props->mime == "image/jpeg") {
-      if (!link_elements(streamer_node, srcfilter, decode, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, decode, convert, props->serial)) return nullptr;
-  } else {
-      if (!link_elements(streamer_node, srcfilter, convert, props->serial)) return nullptr;
-  }
-  
-  if (!link_elements(streamer_node, convert, scalefilter, props->serial)) return nullptr;
-
-  // Enable crop and/or clock
-  if (props->crop43 && props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, encode, props->serial)) return nullptr;
-  } else if (props->crop43) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, encode, props->serial)) return nullptr;
-  } else if (props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, encode, props->serial)) return nullptr;
-  } else {
-      if (!link_elements(streamer_node, scalefilter, encode, props->serial)) return nullptr;
-  }
-
-  if (!link_elements(streamer_node, encode, parse, props->serial)) return nullptr;
-  if (!link_elements(streamer_node, parse, webrtc, props->serial)) return nullptr;
-
-  return gst_pipeline;
-}
-
-/*
- * Retrieve ros2 parameters for h264software pipeline or sets defaults
-*/
-
-h264softwarePipelineProperties* get_h264software_pipeline_properties(rclcpp::Node* streamer_node, camera_msgs::msg::Camera* camera)
-{
-  // 0. Initialize constants
-  h264softwarePipelineProperties* props = new h264softwarePipelineProperties;
-  RCLCPP_DEBUG(streamer_node->get_logger(), "Getting props for %s", camera->serial.c_str());
-  props->serial = camera->serial;
-  props->node = camera->node;
-  props->original_serial = camera->original_serial;
-
-  // Get profile
-  std::string profile = "NULL";
-  streamer_node->get_parameter_or<std::string>((std::string(PIPELINE_PREFIX) + "." + camera->serial + ".profile").c_str(), profile, profile);
-  streamer_node->get_parameter_or<std::string>((std::string(DEFAULT_PREFIX) + "." + camera->original_serial + ".profile").c_str(), profile, profile);
-
-  // 1. Define default properties
-  std::string default_string;
-
-  // source
-  props->device = set_property(streamer_node, camera->serial, profile, camera->original_serial, "device", props->node);
-  default_string = "mmap";
-  props->io_mode = set_property(streamer_node, camera->serial, profile, camera->original_serial, "io_mode", "mmap");
-
-  // filter
-  default_string = "NV12";
-  props->format = set_property(streamer_node, camera->serial, profile, camera->original_serial, "format", default_string);
-  default_string = "image/jpeg";
-  props->mime = set_property(streamer_node, camera->serial, profile, camera->original_serial, "mime", default_string);
-
-  props->brightness = set_property(streamer_node, camera->serial, profile, camera->original_serial, "brightness", 0);
-  props->contrast = set_property(streamer_node, camera->serial, profile, camera->original_serial, "contrast", 0);
-  props->framerate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate", 30);
-  props->framerate_denominator = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate_denominator", 1);
-  props->height = set_property(streamer_node, camera->serial, profile, camera->original_serial, "height", 720);
-  props->width = set_property(streamer_node, camera->serial, profile, camera->original_serial, "width", 1280);
-
-  // decoder
-  default_string = "jpegdec";
-  props->decoder = set_property(streamer_node, camera->serial, profile, camera->original_serial, "decoder", default_string);
-
-  // convert
-  default_string = "nearest";
-  props->chroma_resampler = set_property(streamer_node, camera->serial, profile, camera->original_serial, "chroma_resampler", default_string);
-  default_string = "bayer";
-  props->dither = set_property(streamer_node, camera->serial, profile, camera->original_serial, "dither", default_string);
-  default_string = "nearest-neighbour";
-  props->method = set_property(streamer_node, camera->serial, profile, camera->original_serial, "method", default_string);
-
-  // scale
-  props->downscale = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downscale", 1);
-
-  // rate
-  props->downrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downrate", 1);
-
-  // cropper
-  props->crop43 = set_property(streamer_node, camera->serial, profile, camera->original_serial, "crop43", false);
-
-  // clock
-  props->show_clock = set_property(streamer_node, camera->serial, profile, camera->original_serial, "show_clock", false);
-
-  // encode
-  default_string = "ultrafast";
-  props->speed_preset = set_property(streamer_node, camera->serial, profile, camera->original_serial, "speed_preset", default_string);
-  default_string = "zerolatency";
-  props->tune = set_property(streamer_node, camera->serial, profile, camera->original_serial, "tune", default_string);
-
-  props->gop = set_property(streamer_node, camera->serial, profile, camera->original_serial, "gop", 1);
-  props->noise_reduction = set_property(streamer_node, camera->serial, profile, camera->original_serial, "noise_reduction", 256);
-  props->threads = set_property(streamer_node, camera->serial, profile, camera->original_serial, "threads", 1);
-
-  // webrtc
-  default_string = "gcc";
-  props->congestion_control = set_property(streamer_node, camera->serial, profile, camera->original_serial, "congestion_control", default_string);
-  props->video_caps = "video/x-h264";
-
-  props->bitrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "bitrate", 4096);
-
-  props->do_fec = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_fec", false);
-  props->do_retransmission = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_retransmission", false);
-
-  return props;
-}
-
-/*
- * V4l camera (any) decoded then encoded into vpXenc
- * Enforces alignment from vpX v4l camera and feeds directly to webrtc 
- * gst-launch-1.0 v4l2src device={props->node} ! {props->mime},width={props->width},height={props->height},framerate={props->framerate}/1,alignment={props->alignment},stream-format={props->stream_format},format={props->format}! webrtcsink meta='meta, serial=(string){props->serial}' video-caps=video/x-vp8
- */
-
-GstElement* vpXsoftware_pipeline(rclcpp::Node* streamer_node, vpXsoftwarePipelineProperties* props)
-{
-  // 0. Initialize constants
-
-  // Verify resolution
-  const std::string pipeline_type = "vpXsoftware";
-  if (verify_resolution(props->device, &props->mime, &props->width, &props->height, &props->framerate, &props->framerate_denominator)) {
-      RCLCPP_INFO(streamer_node->get_logger(), "Starting pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  } else {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Wrong resolution! Fallback %s pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  };
-
-
-  // Disable crop43 if it is already 4:3
-  const int crop_width = crop43(props->width, props->height);
-  if (crop_width == 0) {
-      props->crop43 = false;
-  }
-
-  // 1. Create the elements
-  GstElement* gst_pipeline = gst_pipeline_new(props->serial.c_str());
-  GstElement* source = gst_element_factory_make("v4l2src", "video-source");
-  GstElement* rate = (props->downrate > 1) ? gst_element_factory_make("videorate", "rate") : nullptr;
-  GstElement* srcfilter = gst_element_factory_make("capsfilter", "srcfilter");
-  GstElement* decode = (props->mime == "image/jpeg") ? gst_element_factory_make(props->decoder.c_str(), "decoder") : nullptr;
-  GstElement* convert = gst_element_factory_make("videoconvertscale", "converter");
-  GstElement* scalefilter = gst_element_factory_make("capsfilter", "scalefilter");
-  GstElement* encode = (props->video_caps == "video/x-vp9") ? gst_element_factory_make("vp9enc", "encoder") : gst_element_factory_make("vp8enc", "encoder");
-  GstElement* webrtc = gst_element_factory_make("webrtcsink", "webrtc");
-  GstElement* clock = (props->show_clock) ? gst_element_factory_make("clockoverlay", "clock") : nullptr;
-  GstElement* cropper = (props->crop43) ? gst_element_factory_make("videocrop", "video-cropper") : nullptr;
-
-  if (!gst_pipeline || !source || (props->downrate > 1 && !rate) || !srcfilter || (props->mime == "image/jpeg" && !decode) || !convert || !scalefilter  || (props->show_clock && !clock) || (props->crop43 && !cropper) || !encode || !webrtc) {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Could not create pipeline for %s", props->serial.c_str());
-      return nullptr;
-  }
-  
-  // 2. Set element properties
-  set_v4lsource(source, props->device, props->io_mode);
-  set_srcfilter(srcfilter, props->mime, props->width, props->height, props->framerate, props->framerate_denominator, props->downrate, props->brightness, props->contrast);
-  set_convertscale(convert, props->chroma_resampler, props->dither, props->method);
-  set_scalefilter(scalefilter, props->format, props->width, props->height, props->framerate, props->framerate_denominator, props->downscale, props->downrate, props->brightness, props->contrast);
-  set_crop43(cropper, props->crop43, crop_width, props->downscale);
-  set_vpXenc(encode, props->deadline, props->cpu_used, props->end_usage, props->threads, props->bitrate, props->gop, props->framerate, props->framerate_denominator, props->downrate, props->video_caps, props->aq_mode);
-  set_webrtcsink(webrtc, props->serial, props->video_caps, props->do_fec, props->do_retransmission, props->congestion_control, props->bitrate);
-
-  // 3. Add elements to pipeline
-  gst_bin_add_many(GST_BIN(gst_pipeline), source, srcfilter, convert, scalefilter, encode, webrtc, NULL);
-  if (props->crop43) gst_bin_add(GST_BIN(gst_pipeline), cropper);
-  if (props->show_clock) gst_bin_add(GST_BIN(gst_pipeline), clock);
-  if (props->mime == "image/jpeg") gst_bin_add(GST_BIN(gst_pipeline), decode);
-  if (props->downrate > 1) gst_bin_add(GST_BIN(gst_pipeline), rate);
-
-  // 4. Link elements
-  
-  // Change fps
-  if (props->downrate > 1) {
-    if (!link_elements(streamer_node, source, rate, props->serial)) return nullptr;    
-    if (!link_elements(streamer_node, rate, srcfilter, props->serial)) return nullptr;
-
-  } else {
-    if (!link_elements(streamer_node, source, srcfilter, props->serial)) return nullptr;
-  }
-
-  // Convert to raw
-  if (props->mime == "image/jpeg") {
-      if (!link_elements(streamer_node, srcfilter, decode, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, decode, convert, props->serial)) return nullptr;
-  } else {
-      if (!link_elements(streamer_node, srcfilter, convert, props->serial)) return nullptr;
-  }
-  
-  if (!link_elements(streamer_node, convert, scalefilter, props->serial)) return nullptr;
-
-  // Enable crop and/or clock
-  if (props->crop43 && props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, encode, props->serial)) return nullptr;
-  } else if (props->crop43) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, encode, props->serial)) return nullptr;
-  } else if (props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, encode, props->serial)) return nullptr;
-  } else {
-      if (!link_elements(streamer_node, scalefilter, encode, props->serial)) return nullptr;
-  }
-
-  if (!link_elements(streamer_node, encode, webrtc, props->serial)) return nullptr;
-
-  return gst_pipeline;
-}
-
-
-/*
- * Retrieve ros2 parameters for vpXsoftware pipeline or sets defaults
-*/
-
-vpXsoftwarePipelineProperties* get_vpXsoftware_pipeline_properties(rclcpp::Node* streamer_node, camera_msgs::msg::Camera* camera)
-{
-  // 0. Initialize constants
-  vpXsoftwarePipelineProperties* props = new vpXsoftwarePipelineProperties;
-  RCLCPP_DEBUG(streamer_node->get_logger(), "Getting props for %s", camera->serial.c_str());
-  props->serial = camera->serial;
-  props->node = camera->node;
-  props->original_serial = camera->original_serial;
-
-  // Get profile
-  std::string profile = "NULL";
-  streamer_node->get_parameter_or<std::string>((std::string(PIPELINE_PREFIX) + "." + camera->serial + ".profile").c_str(), profile, profile);
-  streamer_node->get_parameter_or<std::string>((std::string(DEFAULT_PREFIX) + "." + camera->original_serial + ".profile").c_str(), profile, profile);
-
-  // 1. Define default properties
-  std::string default_string;
-
-  // source
-  props->device = set_property(streamer_node, camera->serial, profile, camera->original_serial, "device", props->node);
-  default_string = "mmap";
-  props->io_mode = set_property(streamer_node, camera->serial, profile, camera->original_serial, "io_mode", "mmap");
-
-  // filter
-  props->format = "I420";
-  default_string = "image/jpeg";
-  props->mime = set_property(streamer_node, camera->serial, profile, camera->original_serial, "mime", default_string);
-
-  props->brightness = set_property(streamer_node, camera->serial, profile, camera->original_serial, "brightness", 0);
-  props->contrast = set_property(streamer_node, camera->serial, profile, camera->original_serial, "contrast", 0);
-  props->framerate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate", 30);
-  props->framerate_denominator = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate_denominator", 1);
-  props->height = set_property(streamer_node, camera->serial, profile, camera->original_serial, "height", 720);
-  props->width = set_property(streamer_node, camera->serial, profile, camera->original_serial, "width", 1280);
-
-  // decoder
-  default_string = "jpegdec";
-  props->decoder = set_property(streamer_node, camera->serial, profile, camera->original_serial, "decoder", default_string);
-
-  // convert
-  default_string = "nearest";
-  props->chroma_resampler = set_property(streamer_node, camera->serial, profile, camera->original_serial, "chroma_resampler", default_string);
-  default_string = "bayer";
-  props->dither = set_property(streamer_node, camera->serial, profile, camera->original_serial, "dither", default_string);
-  default_string = "nearest-neighbour";
-  props->method = set_property(streamer_node, camera->serial, profile, camera->original_serial, "method", default_string);
-
-  // scale
-  props->downscale = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downscale", 1);
-
-  // rate
-  props->downrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downrate", 1);
-
-  // cropper
-  props->crop43 = set_property(streamer_node, camera->serial, profile, camera->original_serial, "crop43", false);
-
-  // clock
-  props->show_clock = set_property(streamer_node, camera->serial, profile, camera->original_serial, "show_clock", false);
-
-  // encode
-  default_string = "perceptual";
-  props->aq_mode = set_property(streamer_node, camera->serial, profile, camera->original_serial, "aq_mode", default_string);
-  default_string = "cbr";
-  props->end_usage = set_property(streamer_node, camera->serial, profile, camera->original_serial, "end_usage", default_string);
-
-
-  props->cpu_used = set_property(streamer_node, camera->serial, profile, camera->original_serial, "cpu_used", 16);
-  props->deadline = set_property(streamer_node, camera->serial, profile, camera->original_serial, "deadline", 1);
-  props->gop = set_property(streamer_node, camera->serial, profile, camera->original_serial, "gop", 1);
-  props->threads = set_property(streamer_node, camera->serial, profile, camera->original_serial, "threads", 1);
-
-  // webrtc
-  default_string = "gcc";
-  props->congestion_control = set_property(streamer_node, camera->serial, profile, camera->original_serial, "congestion_control", default_string);
-  default_string = "video/x-vp8";
-  props->video_caps = set_property(streamer_node, camera->serial, profile, camera->original_serial, "video_caps", default_string);
-
-  props->bitrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "bitrate", 4096);
-
-  props->do_fec = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_fec", false);
-  props->do_retransmission = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_retransmission", false);
-
-  return props;
-}
-
-/*
- * V4l camera (any) decoded then encoded into av1enc
- * Enforces alignment from av1 v4l camera and feeds directly to webrtc 
- * gst-launch-1.0 v4l2src device={props->node} ! {props->mime},width={props->width},height={props->height},framerate={props->framerate}/1,alignment={props->alignment},stream-format={props->stream_format},format={props->format}! webrtcsink meta='meta, serial=(string){props->serial}' video-caps=video/x-av1
- */
-
-GstElement* av1software_pipeline(rclcpp::Node* streamer_node, av1softwarePipelineProperties* props)
-{
-  // 0. Initialize constants
-
-  // Verify resolution
-  const std::string pipeline_type = "av1software";
-  if (verify_resolution(props->device, &props->mime, &props->width, &props->height, &props->framerate, &props->framerate_denominator)) {
-      RCLCPP_INFO(streamer_node->get_logger(), "Starting pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  } else {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Wrong resolution! Fallback %s pipeline for %s with %dx%d@%dfps", pipeline_type.c_str(), props->serial.c_str(), props->width, props->height, props->framerate/props->framerate_denominator);
-  };
-
-
-  // Disable crop43 if it is already 4:3
-  const int crop_width = crop43(props->width, props->height);
-  if (crop_width == 0) {
-      props->crop43 = false;
-  }
-
-  // 1. Create the elements
-  GstElement* gst_pipeline = gst_pipeline_new(props->serial.c_str());
-  GstElement* source = gst_element_factory_make("v4l2src", "video-source");
-  GstElement* rate = (props->downrate > 1) ? gst_element_factory_make("videorate", "rate") : nullptr;
-  GstElement* srcfilter = gst_element_factory_make("capsfilter", "srcfilter");
-  GstElement* decode = (props->mime == "image/jpeg") ? gst_element_factory_make(props->decoder.c_str(), "decoder") : nullptr;
-  GstElement* convert = gst_element_factory_make("videoconvertscale", "converter");
-  GstElement* scalefilter = gst_element_factory_make("capsfilter", "scalefilter");
-  GstElement* encode = gst_element_factory_make("av1enc", "encoder");
-  GstElement* parse = gst_element_factory_make("av1parse", "parser");
-  GstElement* webrtc = gst_element_factory_make("webrtcsink", "webrtc");
-  GstElement* clock = (props->show_clock) ? gst_element_factory_make("clockoverlay", "clock") : nullptr;
-  GstElement* cropper = (props->crop43) ? gst_element_factory_make("videocrop", "video-cropper") : nullptr;
-
-  if (!gst_pipeline || !source || (props->downrate > 1 && !rate) || !srcfilter || (props->mime == "image/jpeg" && !decode) || !convert || !scalefilter  || (props->show_clock && !clock) || (props->crop43 && !cropper) || !encode || !parse || !webrtc) {
-      RCLCPP_ERROR(streamer_node->get_logger(), "Could not create pipeline for %s", props->serial.c_str());
-      return nullptr;
-  }
-  
-  // 2. Set element properties
-  set_v4lsource(source, props->device, props->io_mode);
-  set_srcfilter(srcfilter, props->mime, props->width, props->height, props->framerate, props->framerate_denominator, props->downrate, props->brightness, props->contrast);
-  set_convertscale(convert, props->chroma_resampler, props->dither, props->method);
-  set_scalefilter(scalefilter, props->format, props->width, props->height, props->framerate, props->framerate_denominator, props->downscale, props->downrate, props->brightness, props->contrast);
-  set_crop43(cropper, props->crop43, crop_width, props->downscale);
-  set_av1enc(encode, props->cpu_used, props->end_usage, props->usage_profile, props->threads, props->bitrate, props->gop, props->framerate, props->framerate_denominator, props->downrate);
-  set_webrtcsink(webrtc, props->serial, props->video_caps, props->do_fec, props->do_retransmission, props->congestion_control, props->bitrate);
-
-  // 3. Add elements to pipeline
-  gst_bin_add_many(GST_BIN(gst_pipeline), source, srcfilter, convert, scalefilter, encode, parse, webrtc, NULL);
-  if (props->crop43) gst_bin_add(GST_BIN(gst_pipeline), cropper);
-  if (props->show_clock) gst_bin_add(GST_BIN(gst_pipeline), clock);
-  if (props->mime == "image/jpeg") gst_bin_add(GST_BIN(gst_pipeline), decode);
-  if (props->downrate > 1) gst_bin_add(GST_BIN(gst_pipeline), rate);
-
-  // 4. Link elements
-  
-  // Change fps
-  if (props->downrate > 1) {
-    if (!link_elements(streamer_node, source, rate, props->serial)) return nullptr;    
-    if (!link_elements(streamer_node, rate, srcfilter, props->serial)) return nullptr;
-
-  } else {
-    if (!link_elements(streamer_node, source, srcfilter, props->serial)) return nullptr;
-  }
-
-  // Convert to raw
-  if (props->mime == "image/jpeg") {
-      if (!link_elements(streamer_node, srcfilter, decode, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, decode, convert, props->serial)) return nullptr;
-  } else {
-      if (!link_elements(streamer_node, srcfilter, convert, props->serial)) return nullptr;
-  }
-  
-  if (!link_elements(streamer_node, convert, scalefilter, props->serial)) return nullptr;
-
-  // Enable crop and/or clock
-  if (props->crop43 && props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, encode, props->serial)) return nullptr;
-  } else if (props->crop43) {
-      if (!link_elements(streamer_node, scalefilter, cropper, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, cropper, encode, props->serial)) return nullptr;
-  } else if (props->show_clock) {
-      if (!link_elements(streamer_node, scalefilter, clock, props->serial)) return nullptr;
-      if (!link_elements(streamer_node, clock, encode, props->serial)) return nullptr;
-  } else {
-      if (!link_elements(streamer_node, scalefilter, encode, props->serial)) return nullptr;
-  }
-
-  if (!link_elements(streamer_node, encode, parse, props->serial)) return nullptr;
-  if (!link_elements(streamer_node, parse, webrtc, props->serial)) return nullptr;
-
-  return gst_pipeline;
-}
-
-
-/*
- * Retrieve ros2 parameters for vpXsoftware pipeline or sets defaults
-*/
-
-av1softwarePipelineProperties* get_av1software_pipeline_properties(rclcpp::Node* streamer_node, camera_msgs::msg::Camera* camera)
-{
-  // 0. Initialize constants
-  av1softwarePipelineProperties* props = new av1softwarePipelineProperties;
-  RCLCPP_DEBUG(streamer_node->get_logger(), "Getting props for %s", camera->serial.c_str());
-  props->serial = camera->serial;
-  props->node = camera->node;
-  props->original_serial = camera->original_serial;
-
-  // Get profile
-  std::string profile = "NULL";
-  streamer_node->get_parameter_or<std::string>((std::string(PIPELINE_PREFIX) + "." + camera->serial + ".profile").c_str(), profile, profile);
-  streamer_node->get_parameter_or<std::string>((std::string(DEFAULT_PREFIX) + "." + camera->original_serial + ".profile").c_str(), profile, profile);
-
-  // 1. Define default properties
-  std::string default_string;
-
-  // source
-  props->device = set_property(streamer_node, camera->serial, profile, camera->original_serial, "device", props->node);
-  default_string = "mmap";
-  props->io_mode = set_property(streamer_node, camera->serial, profile, camera->original_serial, "io_mode", "mmap");
-
-  // filter
-  props->format = "I420";
-  default_string = "image/jpeg";
-  props->mime = set_property(streamer_node, camera->serial, profile, camera->original_serial, "mime", default_string);
-
-  props->brightness = set_property(streamer_node, camera->serial, profile, camera->original_serial, "brightness", 0);
-  props->contrast = set_property(streamer_node, camera->serial, profile, camera->original_serial, "contrast", 0);
-  props->framerate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate", 30);
-  props->framerate_denominator = set_property(streamer_node, camera->serial, profile, camera->original_serial, "framerate_denominator", 1);
-  props->height = set_property(streamer_node, camera->serial, profile, camera->original_serial, "height", 720);
-  props->width = set_property(streamer_node, camera->serial, profile, camera->original_serial, "width", 1280);
-
-  // decoder
-  default_string = "jpegdec";
-  props->decoder = set_property(streamer_node, camera->serial, profile, camera->original_serial, "decoder", default_string);
-
-  // convert
-  default_string = "nearest";
-  props->chroma_resampler = set_property(streamer_node, camera->serial, profile, camera->original_serial, "chroma_resampler", default_string);
-  default_string = "bayer";
-  props->dither = set_property(streamer_node, camera->serial, profile, camera->original_serial, "dither", default_string);
-  default_string = "nearest-neighbour";
-  props->method = set_property(streamer_node, camera->serial, profile, camera->original_serial, "method", default_string);
-
-  // scale
-  props->downscale = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downscale", 1);
-
-  // rate
-  props->downrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "downrate", 1);
-
-  // cropper
-  props->crop43 = set_property(streamer_node, camera->serial, profile, camera->original_serial, "crop43", false);
-
-  // clock
-  props->show_clock = set_property(streamer_node, camera->serial, profile, camera->original_serial, "show_clock", false);
-
-  // encode
-  default_string = "cbr";
-  props->end_usage = set_property(streamer_node, camera->serial, profile, camera->original_serial, "end_usage", default_string);
-  default_string = "realtime";
-  props->usage_profile = set_property(streamer_node, camera->serial, profile, camera->original_serial, "usage_profile", default_string);
-
-  props->cpu_used = set_property(streamer_node, camera->serial, profile, camera->original_serial, "cpu_used", 10);
-  props->gop = set_property(streamer_node, camera->serial, profile, camera->original_serial, "gop", 1);
-  props->threads = set_property(streamer_node, camera->serial, profile, camera->original_serial, "threads", 1);
-
-  // webrtc
-  default_string = "gcc";
-  props->congestion_control = set_property(streamer_node, camera->serial, profile, camera->original_serial, "congestion_control", default_string);
-  props->video_caps = "video/x-av1";
-
-  props->bitrate = set_property(streamer_node, camera->serial, profile, camera->original_serial, "bitrate", 4096);
-
-  props->do_fec = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_fec", false);
-  props->do_retransmission = set_property(streamer_node, camera->serial, profile, camera->original_serial, "do_retransmission", false);
-
-  return props;
-}
-
