+++ src/ros/cameras2/cameras2/docs/research/camera_source_formats/format_dumps/LifeCam_HD-3000.txt
